feat(state-ownership): native-ize `.new` for the QuantHash family (ledger §D ③)
Fourth slice of the ③ built-in type ctor fork. Unlike the prior three,
QuantHash construction (`Set`/`SetHash`/`Bag`/`BagHash`/`Mix`/`MixHash`) is
`&mut self` — element counting plus parameterized-element type checking
(`type_matches_value`) and container-type metadata tagging
(`tag_container_metadata`) — so it cannot live in the static
`try_native_builtin_construct`.

Extract the three `dispatch_new` arms (414 lines) verbatim into a new module
`methods_quanthash_ctor.rs` as the single `&mut self` impl
`try_native_quanthash_construct`. The `dispatch_new` arms become delegations to
it (a true single implementation, no copy — unlike the prior slices' comment-
annotated sibling arms). The VM `.new` dispatch (both the non-mut and mut paths)
calls the new `try_native_quanthash_construct_for_package` wrapper, which strips
a parametric name to its base, gates on `is_quanthash_ctor_type`, and only
clones args when the type matches. Construction is env-pure, so the dispatch is
marked pure. `strip_named_pair_args` widened to `pub(super)`.

VM stats: all six types `new` interpreter fallback → 0; parameterized variants
(`Bag[Int]`, `Set[Int]`) and type-check failures (`X::TypeCheck::Binding`) match
raku. make test 11018; pin t/native-quanthash-ctor.t (18); setbagmix/baggy/mix
whitelist green.
